renamed `cidobj` module to `cidobjrw`
changed to better indicate the read/write primary utility of the objects
