+++ candejar/cidobjrw/cidseq/names.py
