+++ candejar/cidobjrw/cidsubobj_old.py
