+++ candejar/cidobjrw/exc.py
@@ -1,0 +1,8 @@
+# -*- coding: utf-8 -*-
+
+"""`cidobjrw` module exceptions"""
+
+class CIDObjError(Exception):
+    """Base exception for `cidobjrw` module"""
+    pass
+
