+++ tests/test_cidsubobj.py
@@ -1,4 +1,4 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 
-"""Tests for `candejar.cidobj.cidsubobj` module."""
+"""Tests for `candejar.cidobjrw.cidsubobj` module."""
